message d'erreur sur confirmation

+++ js/recap.js
@@ -1,4 +1,14 @@
-console.log(localStorage)
-document.getElementById('orderId').innerHTML = localStorage.orderId;
-document.getElementById('totalPrice').innerHTML = localStorage.totalPrice;
-localStorage.clear()
+console.log(localStorage);
+
+let orderId = localStorage.getItem('orderId');
+
+if (orderId === null) {
+    document.getElementById('recap').style.display = 'none';
+    document.getElementById('oops').style.display = 'block';
+} else {
+    document.getElementById('orderId').innerHTML = localStorage.orderId;
+    document.getElementById('totalPrice').innerHTML = localStorage.totalPrice;
+    localStorage.clear()
+    document.getElementById('recap').style.display = 'block';
+    document.getElementById('oops').style.display = 'none';
+}
